feat(auth): Refactor de utilidades de autentificación y actualizar imports para mejorar consistencia
