+++ src/middleware.ts
@@ -1,7 +1,7 @@
 import { NextResponse } from 'next/server';
 import type { NextRequest } from 'next/server';
 import { jwtVerify } from 'jose';
-import { UserJwtPayload } from './lib/auth';
+import { UserJwtPayload } from './lib/auth-utils';
 
 
 const JWT_SECRET = process.env.JWT_SECRET;
